# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface ProfileInput {
  /** @minLength 2 */
  name?: string;
}

export interface PasswordInput {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface ResumeHistoryItem {
  id: string;
  filename: string;
  uploadedAt: string;
  atsScore: number;
  /** @nullable */
  jobDescription?: string | null;
}

export interface ExtractedInfo {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  skills?: string[];
  education?: string[];
  experience?: string[];
  projects?: string[];
  certifications?: string[];
}

export interface AiAnalysis {
  atsScore?: number;
  missingKeywords?: string[];
  strongSections?: string[];
  weakSections?: string[];
  professionalSummary?: string;
  technicalSkillSuggestions?: string[];
  softSkillSuggestions?: string[];
  jobRoles?: string[];
  interviewQuestions?: string[];
  improvementSuggestions?: string[];
  /** @nullable */
  salaryEstimate?: string | null;
  careerSuggestions?: string[];
  grammarSuggestions?: string[];
  actionVerbSuggestions?: string[];
  keywordOptimizationTips?: string[];
  formattingSuggestions?: string[];
}

export interface AnalysisReport {
  id: string;
  filename: string;
  uploadedAt: string;
  atsScore: number;
  /** @nullable */
  jobDescription?: string | null;
  /** @nullable */
  resumeText?: string | null;
  extractedInfo?: ExtractedInfo;
  aiAnalysis?: AiAnalysis;
}

export type DashboardStatsScoreDistributionItem = {
  range: string;
  count: number;
};

export interface DashboardStats {
  totalAnalyses: number;
  averageAtsScore: number;
  highestAtsScore: number;
  recentAnalyses: ResumeHistoryItem[];
  scoreDistribution?: DashboardStatsScoreDistributionItem[];
}