# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Resume Analyzer API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: resumes
    description: Resume management
  - name: dashboard
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /auth/password:
    put:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/account:
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Delete user account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /resumes:
    get:
      operationId: getResumeHistory
      tags: [resumes]
      summary: Get user resume history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Resume history list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResumeHistoryItem"

  /resumes/{id}:
    get:
      operationId: getAnalysis
      tags: [resumes]
      summary: Get analysis for a specific resume
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume and its analysis
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    UserProfile:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"

    ProfileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2

    PasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6

    ResumeHistoryItem:
      type: object
      required: [id, filename, uploadedAt, atsScore]
      properties:
        id:
          type: string
        filename:
          type: string
        uploadedAt:
          type: string
          format: date-time
        atsScore:
          type: integer
        jobDescription:
          type: ["string", "null"]

    ExtractedInfo:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string

    AiAnalysis:
      type: object
      properties:
        atsScore:
          type: integer
        missingKeywords:
          type: array
          items:
            type: string
        strongSections:
          type: array
          items:
            type: string
        weakSections:
          type: array
          items:
            type: string
        professionalSummary:
          type: string
        technicalSkillSuggestions:
          type: array
          items:
            type: string
        softSkillSuggestions:
          type: array
          items:
            type: string
        jobRoles:
          type: array
          items:
            type: string
        interviewQuestions:
          type: array
          items:
            type: string
        improvementSuggestions:
          type: array
          items:
            type: string
        salaryEstimate:
          type: ["string", "null"]
        careerSuggestions:
          type: array
          items:
            type: string
        grammarSuggestions:
          type: array
          items:
            type: string
        actionVerbSuggestions:
          type: array
          items:
            type: string
        keywordOptimizationTips:
          type: array
          items:
            type: string
        formattingSuggestions:
          type: array
          items:
            type: string

    AnalysisReport:
      type: object
      required: [id, filename, uploadedAt, atsScore]
      properties:
        id:
          type: string
        filename:
          type: string
        uploadedAt:
          type: string
          format: date-time
        atsScore:
          type: integer
        jobDescription:
          type: ["string", "null"]
        resumeText:
          type: ["string", "null"]
        extractedInfo:
          $ref: "#/components/schemas/ExtractedInfo"
        aiAnalysis:
          $ref: "#/components/schemas/AiAnalysis"

    DashboardStats:
      type: object
      required: [totalAnalyses, averageAtsScore, highestAtsScore, recentAnalyses]
      properties:
        totalAnalyses:
          type: integer
        averageAtsScore:
          type: number
        highestAtsScore:
          type: integer
        recentAnalyses:
          type: array
          items:
            $ref: "#/components/schemas/ResumeHistoryItem"
        scoreDistribution:
          type: array
          items:
            type: object
            required: [range, count]
            properties:
              range:
                type: string
              count:
                type: integer