# lib/api-zod/src/generated/types/aiAnalysis.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface AiAnalysis {
  atsScore?: number;
  missingKeywords?: string[];
  strongSections?: string[];
  weakSections?: string[];
  professionalSummary?: string;
  technicalSkillSuggestions?: string[];
  softSkillSuggestions?: string[];
  jobRoles?: string[];
  interviewQuestions?: string[];
  improvementSuggestions?: string[];
  /** @nullable */
  salaryEstimate?: string | null;
  careerSuggestions?: string[];
  grammarSuggestions?: string[];
  actionVerbSuggestions?: string[];
  keywordOptimizationTips?: string[];
  formattingSuggestions?: string[];
}