# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsScoreDistributionItem } from './dashboardStatsScoreDistributionItem';
import type { ResumeHistoryItem } from './resumeHistoryItem';

export interface DashboardStats {
  totalAnalyses: number;
  averageAtsScore: number;
  highestAtsScore: number;
  recentAnalyses: ResumeHistoryItem[];
  scoreDistribution?: DashboardStatsScoreDistributionItem[];
}