# lib/api-zod/src/generated/types/resumeHistoryItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeHistoryItem {
  id: string;
  filename: string;
  uploadedAt: Date;
  atsScore: number;
  /** @nullable */
  jobDescription?: string | null;
}