# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  createdAt: Date;
}